Add context around ParameterCollection._ArmiObject
I spent a while trying to figure out why the armi object attached to a parameter
collection was not the same as the object that held the parameter
collection. Or, why `self.p._ArmiObject is not self` evaluates to False
but `isinstance(self, self.p._ArmiObject)` does evalute to True. The
comment attached to this attribute mentions it is a class but so much
of armi revolves around this `node.parent` tree-like behavior that I felt
or wanted to have this tree-behavior extend to parameter collections.

This would make some work on the TD fraction easier because the
parameter setter would have easier access to the `ArmiObject`
**instance** that held that collection.