Provide Component.finalizeLoadingFromDB to set Material TD from DB
This exists because the theoretical density is initially defined as a material
modification, and then stored as a Material attribute. When reading from blueprints,
the blueprint loader sets the theoretical density parameter from the Material
attribute. Component parameters are also set when reading from the database.
But, we need to set the Material attribute so routines that fetch a material's
density property account for the theoretical density.

I do not like this solution. But I could not find a clean way to have the
setting of a parameter do something with the object that stores that
parameter. My original thought was something like
```python
def _assignTDFrac(self, val):
    self._ArmiObject.material.adjustTD(val)
    self._p_theoreticalDensityFrac = val
```

but the `ArmiObject` is not the "parent" of the parameter collection.

I also tried to find a way to use existing `Component` methods that may
be called during construction and loading from database, but I couldn't
find a good fit. The process appears to be

1. Create the material by string name (this loses material modifications
   and instance attributes that existed when written to DB)
2. Create a fresh instance of the component-type (e.g., `Circle`)
3. Later, assign all parameters found in the database to the new
   component

The material is already created and attached to the component by the
time parameters are read in.

Call Component.finalizeLoadingFromDB when composing reactor from DB