Changing the data type of thKernel.
The thKernel setting is meant to store the name of the primary
thermal hydraulics solver for a run. The default value was False,
causing the data type of thKernel to be bool. When a user set the
value of thKernel to anything in the settings file, it would only
be set to True. Changing the default value to an empty string
changes its data type to str, allowing the setting to work as
originally intended.