Track down where Material.density is called more exactly
Using traceback, we can find the thing that called `Material.density`
when the material is attached to a `Component`. This information,
specifically the file name and line number, is used in the warning
message. The file name and line number is used as the label in the
warning so multiple locations that invoke this behavior will show up
as one logged message. But their number of invocations will also be
counted.