Update jaggedArray to read/store empty lists
Empty lists could be stored in the database under the old
jagged array implementation. The new implementation in
PR #1726 treated all empty arrays as None, and was incapable
of reading empty lists from old databases that had them.

This PR updates the packing and unpacking routines for special data
so that empty lists can be stored and read back from the database.