ExpansionData.determineTargetComponent returns target
Useful for testing what you get is what you expect. Tests added
to this effect.

Renamed an internal variable from `componentWFlag` to `candidates` because
the name implied `componentWFlag` was a single component when it was a
(potentially empty) list of components.

Finally, bring out the setting of the target component to a separate
method. This is useful for subclasses that may need custom logic to
determine their target component, but want to perform the same actions
on the setter.