Provide AssemblyAxialLinkage.getLinkedBlocks
This implementation an O(N^2) iteration where each block also iterates over all blocks in the assembly. The use of itertools.chain and itertools.islice provide
a more efficient and streamlined iteration look.

The call signature is for a `Sequence[Block]` which could just as easily be
`Assembly`. We can't use just `Iterable[Block]` because we need to perform
multiple iterations on the object at once. And things like a generator may
get consumed part way through the iteration. And for testing, passing a list
of blocks gets the same behavior as passing one assembly.

Tests for zero blocks, one block, and multiple blocks added.