Add ENDF/B-VII.1 MC2-3 ID dictionary
To support the VII.1 MC2-3 library, ARMI needs a lookup dictionary by
library ID. Previously, there was just a single MC2-3 dictionary
to do this, but now we need two - one for the VII.0 library and one
for the VII.1 library. The preexisting `byMCC3Id` dictionary is still
availabe and now returns VII.1 data.

Also:
* adds getter functions for the VII.0 and VII.1 IDs on a couple classes
* modifies existing versionless getMcc3ID getters to wrap VII.1 getters
* modifies readMCCNuclideData to parse the VII.1 nuclide labels