Use dict of ij to powers in optimal assembmly orientation
The previous list approach, using the index method, got a little finicky
depending on how `IndexLocation.__eq__` is implemented and on who's grid
the location returned from `HexGrid.rotateIndex` belongs. So now just
make a dictionary mapping `{(i, j): p}` for each location in the
previous block.

We could go a little more optimal and only store locations if they're in
the same ring as the pin with the most burnup. But then we have to pass
each location through `HexGrid.getRingPos`. And, at most we have maybe a
few dozen entries per ring vs maybe 200 in a full hexagon. Splitting
hairs but it may be useful to keep in mind.