buReducingAssemblyRotation handles fresh assemblies better
In some testing, I found that freshly created assemblies caused problems
with the burnup reducing rotation algorithm. They didn't have a useful
previous location, in that their `lastLocationLabel` was `LoadQueue`
since they had yet to be placed within the core. This caused problems
trying to find the assembly that now exists where they used to be
because the logic for breaking an assembly locator label like `001-003`
breaks when that label is `LoadQueue`.

The fuel handler tests now work higher up on the `FuelHandler.outage`
method that invokes the `buReducingAssemblyRotation` function.

In the case of the fresh fuel, we fake a case where an assembly is
created fresh and dropped into the reactor without moving another
assembly. `FuelHandler.moved` only really makes sense for `N>1` entries
if you're swapping or disharging assemblies. But, we want to show that
we skip assemblies that are coming from outside the reactor.