Remove need for history interface in burnup equalizing rotation
It seems like checking if an assembly was in the detailed assembly prior
to rotation was a means to check if it had pin power. But the workflow
for adding detailed assemblies seems less widespread compared to setting
pin powers. So, we'll first check any block in the assembly being
rotated has pin power prior to the detailed assembly check.

Check that the current assembly, `aNow`, has burnup defined on pins.
Additionally check that the assembly previously located here, `aPrev`,
has power defined on pins.

These checks are facilitated by moving checks for assembly fuel pin
powers and burnups into `fuelCycle/utils.py`