Experimental: check assembly pin power prior to rotation
It seems like checking if an assembly was in the detailed assembly prior
to rotation was a means to check if it had pin power. But the workflow
for adding detailed assemblies seems less widespread compared to setting
pin powers. So, we'll first check any block in the assembly being
rotated has pin power prior to the detailed assembly check.