Extending axial expansion classes (#1920)
* Flatten and rename determine axial linkage function

Use early `return` statements to dedent the function. Renamed
to `areAxiallyLinked` to be a bit more declarative, even though
this function lives in the axial expansion routine.

* Provide axialAssemblyLinkage.AxialLink for improved readability

The values in the linkage dictionaries on `AssemblyAxialLinkage` were
lists of two items where the position meant something: `0` for linked
item below the current thing, and `1` for linked item above the current
thing. Some functionality was based on `links[1] is not None` which is
not super readable unless you're familiar with what's going on.

This introduces a small data class with two attributes: `lower` and
`upper` such that `links.lower is links[0]` and `links.upper is
links[1]`. This improves readability and extensibility by making a more
explicit and declarative thing. The position-based getters and setters
are retained and tested.

The class is "templated" with a `typing.Generic[typing.TypeVar]` that
could be either `Block` or `Component`. This is because the the two
linked dictionaries on the `AssemblyAxialLinkage` class only differ
based on the types of their values. We still need to know lower/upper as
determined by the values, but for the block link dict we want blocks and
then components from the component link dict.