Don't rotate assemblies until all assembly rotations have been determined
If we rotate an assembly that is then used to determine how a different
assembly should be rotated, we've poisoned our check. We want to compare
against the state of the assembly that was in the core when the power
was produced.

Consider:

1. Assembly A is shuffled to a new location and then rotated.
2. Assembly B is shuffled to the previous location of assembly A.

The power profile that is used to determine the rotation of assembly B
is based on a rotated assembly A. This is not ideal because we want to
compare against the power profile assembly B is expected to experience.
And by rotating assembly A before this check, we've introduced a bias
into the check.

Now, the `buReducingAssemblyRotation` will store the number of rotations
for each assembly first, then rotate all assemblies at once.

Additionally, any assembly that is found to have a zero rotation is not
going to be rotated. Calling `Assembly.rotate(0)` is a no-op that we
don't need to consider. These no-rotations will also be excluded from
the counting of rotations.