fix unit test and example notebook
the unit test was comparing the power in a block between two timesteps.
the block was being moved from a non-central location to the central
location, where only 1/3rd of the assembly should be in the reactor. The
assert in the test was incorrect in that it was saying the power in the
assembly should be the same before and after the move. This is wrong for
a few reasons, first being that the power is no longer valid after the
move. But aside from that, the power should be 1/3rd what it was because
the assembly is only 1/3rd in the core. Really it should be zeroed out
but other non-power parameters that are volume calculated should be
reduced by the amount that the assembly actually exists in the reactor.

Additionally, the power in the central assembly in the example was
calculated assuming the full assembly exists in the core. the example
was modified to correctly calculate the power in the central assembly,
which then required FuelHandler._getParamMax to account for the symmetry
factor of the assembly, otherwise the ordering was incorrect and the
shuffling in the example occured in a different order.