Provide Core.iterBlocks
After a survey of some applications, there were a lot of patterns like
```python
for b in r.core.getBlocks():
```
which involves traversing the reactor, populating a list of blocks, and
then returning that list for iteration.

This change introduces a method that produces an iterator of all blocks
in a core, optionally with some flag and/or satisfying a condition. It
relies on `Composite.iterChildren` with the assumption that the
composite tree flows `Core` -> `Assembly` -> `Block`.

Does not support the full level of selection that `Core.getBlocks`
supports. `Core.getBlocks`, which relies on `Core.getAssemblies` allows
for selecting blocks that do not exist in the `Core`; users can also
select blocks from the spent fuel pool and blocks from blueprints.