Use a looser tolerance on pitch for HexGrid roughly equal check
We don't need a super strict check to make sure the grids are roughly
equal. The ideal of the method is to check that one index location in
one grid would be in approximately the same spatial location in the
other grid. Otherwise we can't accurately rotate the position within the
grid.

This problem came up in some shuffling work where two assemblies had
ever so slightly different pitches, on the order of hundreths of a
percent. The strict `other.pitch == self.pitch` failed even though, for
the purposes of this method, the two grids are "roughly equal"