Ignore Fluid density warning for fluids attached to components
https://github.com/terrapower/armi/pull/1852 introduced a warning
when `Component.material.density(...)` was invoked. The rationale is
the `Component` is the source of truth for volume, density, mass, etc.
And components are the things that are thermally expanded.

However, fluids are not thermally expanded and so the density of a fluid
material is appropriate.

This change undoes the density wrapper applied by the base `Material`
when creating a `Fluid` subclass.

A test is added to show that any non-`Fluid` subclass still logs the
warning, but a `Fluid` subclass does not.

I attempted to use a "real" material like `Sodium` but lots of other
warnings can be logged depending on appropriate bounds of temperatures.
So a fake Fluid is used to have more control over what warnings are
logged.