Use np.int32 when reading geodst files
A user encountered a problem reading a GEODST file with at least
32773 indices along one axis of the GEODST matrix. This caused an
`OverflowError` by numpy trying to downcast `32773` to a short integer
with `np.int16`.

We could just increase the size and call it good. But this led us down a
small rabbit hole of "is ARMI missing the spec or trying to be clever?"
And I think ARMI missed the spec here and the `coarseMeshRegions` should
be a "long int" or `np.int32`.

The GEODST specification [1] lists this field as a integer through the
magic implicit Fortran `M` variable naming. So this tells us it
should match the default fortran integer sizing of 8-byte or (c-speak)
`long int` or `np.int32` [2].

Sources:

[1] https://www.osti.gov/servlets/purl/4350057 - CCCC specification from
1974, photo scanned so quality is not great. See the 6D record on page
33 for "Region Assignments to Coarse Mesh Intervals"

[2] https://docs.oracle.com/cd/E19957-01/802-2998/802-2998.pdf - Fortran 77
specification, integer constants on page 30. Note the integer range
by default is `(-2,147,483,648 to 2,147,483,648)` or a 8-byte integer.
A short integer (page 31) has an upper limit range 32,767 which is lower
than the size the model that triggered this investigation.