Initial approach for Component.p.pinIndices
Provide Block.assignPinIndices to be called during the initial
orientation of a block when the spatial grid is automatically created
or is already present because of a lattice grid input.

The idea is, for a block with `locations = b.getPinLocations()` as a
list of `IndexLocations` for all the pins in the block, circular
components have a `Component.p.pinIndices` parameter. This data is used
to determine where each of the pins reside in space, and where data for
the pins reside in various data arrays like `linPowByPin`. For example,
```python
locations = b.getPinLocations()
fuel = b.getComponents(Flags.FUEL)[0]
fuelLocations = list(fuel.spatialLocator)
assert len(fuelLocations) == fuel.p.pinIndices.size
for ix, loc in zip(fuel.p.pinIndices, fuelLocations):
    assert loc == locations[ix]
```
One could then find all the pin-powers with
```python
pinPows = b.p.linPowByPin
powersForThisPin = pinPows[fuel.p.pinIndices]
```