Making ParamMapper symmetry-aware (#2208)
* make ParamMapper convert parameter values to full size

The parameter mapper did not account for symmetry when mapping
parameters between blocks. This was an issue when using the parameter
mapper to assign parameters to an assembly that was not yet placed in
the reactor grid from an assembly that had a non-unity symmetry factor.
Now, the ParamMapper will internally convert volume integrated
parameters to full-block values when getting parameters, and reduce
parameters to partial values when assigning.