Make DimensionLink public
Used mainly in `component.py` but also in `cases` so there's _some_
evidence it can be public. Needed because I want to have the
`pinIndices` parameter be a linked dimension but it's up to the block to
set that. So we need this to be used in at least `blocks.py`