pinIndices either numpy array or None
None no longer signifies this could be a pin that fully occupies the
lattice site. And the decision to find the "linked" pin is now entirely
performed at invocation time, not through a dimension link.

The problem was hard to handle how to serialize the data array that
could be
1. None,
2. array of ushort,
3. DimensionLink

The JaggedArray behavior did not support this, partially becuase there
was no consistent data type between the entries. A similar problem was
encountered if the "linked" pin was stored just by name.

Now we check to find the linked pin based on the equality of the spatial
locators during the `Component.getPinIndices` call. Rather than the
`Block.assignPinIndices` call at construction. I don't love it. But, we
_should_ really only be calling the `getPinIndices` for fuel or control
pins where we have that data immediately available.