Re-assign pin indices when sorting the block
The component pin indices are assigned when being added to an
assembly. This is done well in advance of the reactor being
completed, e.g., in `reactors.factory()`. One of the last things
the factory method does is call `Reactor.sort` which will eventually
call `Block.sort` and change the order of the child components in the
block. This, in turn, changes the pin locations since now we could
find clad components in different order. Which would render
`Component.getPinIndices` meaningless because the locations are relative
to the previous ordering of components.

Now, `Block.sort` will re-assign pin indices if there is a spatial grid.