Update block level BOL HM through axial expansion
Axial expansion changes component parameters lik molesHmBOL. But
it did not change the block level parameters to reflect this. That means
you could have a block molesHmBOL that did not equal the sum of that
value across the child components.

This adds a test demonstrating failed behavior and a change to correctly
set the block molesHmBOL and massHmBOL