Pass around locator type during db load
This helps us know what type of locator we've got. For index locations
and multi-index locations, we can directly query the grid[i,j,k] and get
the correct IndexLocation / MultiIndexLocation object out. But, if we
pass the indices of the coordinate location in, we get an index location
out. And that's problematic because we don't get a 1:1 identical reactor
on the reload.

We dump the location type information to DB already. This changes the
layout `_initComps` to produce that information in sync w/ the location
information. Then we can do a check if we're looking at a coordinate
location or not. This means if a coordinate location goes in, a
coordinate location comes out of the database.