prepRestart hook and Operator._onRestart
New method `Operator._onRestart` is added to handle and delegate the
work at the start of a restart run.

Operator calls this after interactInit but before interactBOL for
restart runs. It copies some of the code from the main interface
interactBOL logic and splits it in two parts.

First, the bookkeeping plugin will find the database interface
and call `prepRestartRun`. This will load the previous state and merge
histories.

If we're re-loading at the start of a new cycle, the operator will call
interactAllEOC. This exists primarily to support shuffling that takes
place during the EOC hook.

The bookkeeping plugin and datbase `prepRestartRun` is handled through a
new plugin hook: `prepRestart`. The bookkeeping plugin attempts to go
first since it needs to load the previous state and attach the previous
reactor in memory. Other plugins can go after.